Show rendered banner if only disabled
When you add and configure a template for a banner, you see it with the last saved information in the overview. When you disabled it, an empty template was shown and you had to open its configuration in order to see, what is configured for it. With this change, you will still see the preview with the last saved text, so that you don't have to open its configuration first.