Automatically wrap too long text on templates
If you have text with line breaks or which need line breaks on your templates, you have to add each line to an extra configuration with new X-Y-Coordinates and font settings.

This change improves this by automatically wrapping too long text, which would not fit on the template without any line break. This is automatically done based on the font size, text length, X-Coordinate and template width.