Fix/Improve error `nickname is already in use`
When the TeamSpeak bot is currently connected and you tried to edit the instance, you have most of the time seen a "HTTP 500 Server Error" page as the clients nickname was already in use by the bot. This change handles this error case and implements a better connection retry logic.